requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pandas>=2.2.0
numpy==1.26.3
scikit-learn==1.3.2
catboost==1.2.2
mlforecast==0.11.3
mlflow==2.10.0
streamlit==1.30.0
plotly==5.18.0
requests==2.31.0
huggingface_hub==0.20.3
entsoe-py==0.6.3
utilsforecast
joblib
google-cloud-storage
gcsfs

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
pydantic==2.6.0